+++ src/main/java/cn/mcmod/tsuki/Tsuki.java
@@ -9,8 +9,8 @@ import cn.mcmod.tsuki.client.particle.ParticleRegistry;
 import cn.mcmod.tsuki.init.CreativeModeTabRegistry;
 import cn.mcmod.tsuki.init.EntityTypeRegistry;
 import cn.mcmod.tsuki.init.ItemDataComponentRegistry;
-import cn.mcmod.tsuki.init.MenuTypeRegistry;
 import cn.mcmod.tsuki.init.LootModifierRegistry;
+import cn.mcmod.tsuki.init.MenuTypeRegistry;
 import cn.mcmod.tsuki.init.RecipeTypeRegistry;
 import cn.mcmod.tsuki.init.SoundEventRegistry;
 import cn.mcmod.tsuki.init.TreeDecoratorTypeRegistry;
@@ -89,13 +89,13 @@ public class Tsuki {
         MobEffectRegistry.MOB_EFFECTS.register(modEventBus);
         RecipeTypeRegistry.RECIPE_TYPES.register(modEventBus);
         RecipeTypeRegistry.RECIPE_SERIALIZERS.register(modEventBus);
-        LootModifierRegistry.LOOT_MODIFIER_SERIALIZERS.register(modEventBus);
         FeatureTypeRegistry.FEATURES.register(modEventBus);
         TreeDecoratorTypeRegistry.TREE_DECORATOR_TYPES.register(modEventBus);
         VillagerRegistry.POI_TYPES.register(modEventBus);
         VillagerRegistry.PROFESSIONS.register(modEventBus);
         CreativeModeTabRegistry.TABS.register(modEventBus);
         ItemDataComponentRegistry.DATA_COMPONENTS.register(modEventBus);
+        LootModifierRegistry.LOOT_MODIFIER_SERIALIZERS.register(modEventBus);
 
         if (ModList.get().isLoaded("guideme")) {
             TsukiGuideMeCompat.register();