+++ src/main/java/cn/mcmod/tsuki/init/LootModifierRegistry.java
@@ -8,10 +8,13 @@ import net.neoforged.neoforge.registries.DeferredHolder;
 import net.neoforged.neoforge.registries.DeferredRegister;
 import net.neoforged.neoforge.registries.NeoForgeRegistries;
 
-public class LootModifierRegistry {
+public final class LootModifierRegistry {
     public static final DeferredRegister<MapCodec<? extends IGlobalLootModifier>> LOOT_MODIFIER_SERIALIZERS = DeferredRegister
             .create(NeoForgeRegistries.Keys.GLOBAL_LOOT_MODIFIER_SERIALIZERS, Tsuki.MODID);
 
     public static final DeferredHolder<MapCodec<? extends IGlobalLootModifier>, MapCodec<FishingCatchLootModifier>> FISHING_CATCH = LOOT_MODIFIER_SERIALIZERS
             .register("fishing_catch", () -> FishingCatchLootModifier.CODEC);
+
+    private LootModifierRegistry() {
+    }
 }